pyproject.toml
[project]
name = "credora"
version = "0.1.0"
description = "AI-driven CFO platform for e-commerce businesses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.6.4",
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.100.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hypothesis]
max_examples = 100
